requirements.txt
opencv-python
mediapipe
numpy
tensorflow
pyautogui
Pillow
tk
